fix: corrected errors from tests
